Changes made to Ublox Driver as well as I2C task. Accelerometer needs changes to make it more stable and less sensitive to small movements

+++ Components/I2C/I2C_Task.hpp
@@ -35,7 +35,7 @@ public:
     /**Getters for CAN formatting */
     uint8_t GPSData();
     uint8_t* PollData();
-    uint8_t CheckSum();
+    uint16_t CheckSum();
 
     static constexpr uint8_t NAV_PVT_DATA =  0x07;  // same thing as the UBX class=0x01,msg=0x07
